break up graph rag into smaller scripts
