+++ example/graph_rag/README.md
@@ -1,0 +1,4 @@
+
+# Graph RAG
+
+Taken from python notebook on a [branch](https://github.com/ibm-granite-community/granite-legal-cookbook/blob/158-legal-graph-rag/recipes/Graph/Entity_Extraction_from_NH_Caselaw.ipynb)
